server/requirements.txt
fastapi
uvicorn[standard]
python-socketio
fastapi-socketio
pydantic
pytest
pytest-asyncio
httpx